requirements.txt
jupyterlab
matplotlib
numpy
pandas
pillow